refactor: migrate fuzz tests from fuzz_test.go to config_test.go for better organization
